https://data.worldbank.org/indicator/SE.TER.ENRR?view=chart
School enrollment, tertiary (% gross)
Gross enrollment ratio is the ratio of total enrollment, regardless of age, to the population of the age group that officially corresponds to the level of education shown. Tertiary education, whether or not to an advanced research qualification, normally requires, as a minimum condition of admission, the successful completion of education at the secondary level.