https://data.worldbank.org/indicator/SI.POV.DDAY?view=chart
Poverty headcount ratio at $1.90 a day (2011 PPP) (% of population)
Poverty headcount ratio at $1.90 a day is the percentage of the population living on less than $1.90 a day at 2011 international prices. As a result of revisions in PPP exchange rates, poverty rates for individual countries cannot be compared with poverty rates reported in earlier editions.